refactor: 메달 종류 Enum 대회 도메인으로 변경
