+++ src/main/java/umc/cockple/demo/domain/contest/enums/MedalType.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.contest.enums;
 
 public enum MedalType {
